requirements.txt
requests==2.31.0
openai==1.35.0
pandas==2.2.2
sqlalchemy==2.0.30
streamlit==1.35.0
python-dotenv==1.0.1
transformers==4.41.2
torch==2.3.0
plotly==5.22.0
